add option to parse raw IP packets
The tool dnscap creates pcap files that do not contain any ethernet
frames but start directly with the IPv4 or IPv6 packets.

This is causing problems because the parser stops when it finds a
packet that doesn't match the starting point.

To parse these files this commit adds a new LayerType which detects if
the family is IPv4 or IPv6 and then tells the parser where to continue.

The old behaviour is kept and the new way is kept behind a switch
`--noEthernetframe`.